Share Podman image transfer test fixture
