+++ internal/deploy/podman/image_transfer_pipe_test.go
@@ -2,10 +2,7 @@ package podman_test
 
 import (
 	"context"
-	"fmt"
-	"path/filepath"
 	"testing"
-	"time"
 
 	"github.com/arm/topo/internal/deploy/podman"
 	"github.com/arm/topo/internal/ssh"
@@ -38,24 +35,4 @@ func assertImageExists(t *testing.T, socket podman.Socket, imageName string) {
 	t.Helper()
 	err := podman.Command(t.Context(), socket, "image", "exists", imageName).Run()
 	assert.NoError(t, err)
-}
-
-func imageTransferFixture(t *testing.T) (string, string) {
-	t.Helper()
-	temporaryDirectory := t.TempDir()
-	composeFile := filepath.Join(temporaryDirectory, "compose.yaml")
-	imageName := "test-image-" + sanitiseTestName(t)
-	requireWriteFile(t, composeFile, fmt.Sprintf(`
-services:
-  test:
-    build: .
-    image: %s
-`, imageName))
-	requireWriteFile(t, filepath.Join(temporaryDirectory, "Dockerfile"), "FROM docker.io/library/alpine:latest\n")
-	t.Cleanup(func() {
-		ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-		defer cancel()
-		_ = podman.Command(ctx, podman.LocalSocket, "image", "rm", "-f", imageName).Run()
-	})
-	return composeFile, imageName
 }
