+++ internal/deploy/podman/testutil_test.go
@@ -1,8 +1,13 @@
 package podman_test
 
 import (
+	"context"
+	"fmt"
+	"path/filepath"
 	"testing"
+	"time"
 
+	"github.com/arm/topo/internal/deploy/podman"
 	gtestutil "github.com/arm/topo/internal/testutil"
 )
 
@@ -24,4 +29,24 @@ func requireWriteFile(t *testing.T, path, content string) {
 func requireAvailableTCPPort(t *testing.T) string {
 	t.Helper()
 	return gtestutil.RequireAvailableTCPPort(t, "127.0.0.1")
+}
+
+func imageTransferFixture(t *testing.T) (string, string) {
+	t.Helper()
+	temporaryDirectory := t.TempDir()
+	composeFile := filepath.Join(temporaryDirectory, "compose.yaml")
+	imageName := "test-image-" + sanitiseTestName(t)
+	requireWriteFile(t, composeFile, fmt.Sprintf(`
+services:
+  test:
+    build: .
+    image: %s
+`, imageName))
+	requireWriteFile(t, filepath.Join(temporaryDirectory, "Dockerfile"), "FROM docker.io/library/alpine:latest\n")
+	t.Cleanup(func() {
+		ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
+		defer cancel()
+		_ = podman.Command(ctx, podman.LocalSocket, "image", "rm", "-f", imageName).Run()
+	})
+	return composeFile, imageName
 }
